Simplified, removed Permissive Policy
