+++ functions.Dockerfile
@@ -2,6 +2,12 @@ FROM mcr.microsoft.com/azure-functions/python:3.0-python3.9-buildenv as build
 ARG BUILD_BRANCH=""
 ARG BUILD_SHA=""
 
+# Remove jessie from sources.list, it's EOL
+RUN sed -i '/jessie/d' /etc/apt/sources.list
+
+# Upgrade any OS packages with outstanding upgrades, to ensure we've got security fixes.
+RUN apt-get update && apt-get upgrade -qyy && rm -rf /var/lib/apt/lists/*
+
 WORKDIR output
 COPY src/requirements.txt .
 RUN pip install --target=./ -r ./requirements.txt
@@ -9,6 +15,15 @@ COPY src/ .
 RUN echo "${BUILD_BRANCH}\n${BUILD_SHA}" > "/output/build_info.txt"
 
 FROM mcr.microsoft.com/azure-functions/python:3.0-python3.9-slim
+
+# Remove jessie from sources.list, it's EOL
+RUN sed -i '/jessie/d' /etc/apt/sources.list
+
+# Upgrade any OS packages with outstanding upgrades, to ensure we've got security fixes.
+RUN apt-get update && apt-get upgrade -qyy && rm -rf /var/lib/apt/lists/*
+
+# Remove jessie from sources.list, it's EOL
+RUN sed -i '/jessie/d' /etc/apt/sources.list
 
 ENV \
     # Enable detection of running in a container
